requirements.txt
Django==2.2.5
djangorestframework==3.10.3
coverage==4.5.4
gunicorn==19.9.0
requests>=2.20
django-cors
django-cors-headers==3.2.0